// DesignAutomator/ClientApp/src/app/models/dna/filtereduserdna.ts
export class FilteredUserDna {
  isDefault: string;
  clientId: number;
  dnaClientId:number;
  uname: string;
  dnalogo: string;
  colour1: string;
  colour2: string;
  colour3: string;
  colour4: string;
  colour5: string;
  naming: string;
  clientName: string;
  clientAddress1: string;
  clientAddress2: string;
  clientCountry: string;
  clientPostCode: string;
}